starter/requirements.txt
azure-keyvault-secrets
azure-identity
azure-storage-blob
azure-ai-contentsafety
langchain
langchain-core
langchain-community
langchain_openai
langchain-chroma
langgraph
pandas
numpy
pydantic
pymongo
bson
fairlearn
presidio-analyzer
psycopg[binary]
sqlalchemy
PyPDF2
torch
transformers
Pillow
matplotlib
